Fix --list --ignored behavior
I'm trying to get libtest-mimic to work well with
[cargo-nextest](https://nexte.st/) -- see the documentation on [custom
test harnesses](https://nexte.st/book/custom-test-harnesses.html) and
https://github.com/nextest-rs/nextest/issues/38. I'd love to recommend
libtest-mimic as a way to write custom test harnesses.

One of the requirements for nextest is the ability to detect ignored and
non-ignored tests during the list phase. libtest prints out just the
ignored tests if you run `--list --ignored`: make libtest-mimic do the
same.

In my testing, this appears to be the only issue blocking libtest-mimic
from working with nextest.