Replace `rayon` with `threadpool` and use main thread for -j1
Ditching rayon and crossbeam_channel gets rid of quite a few
(indirect) dependencies. For this use case, we don't need anything
special at all. We special case `--test-threads=1` to not use a thread
pool at all and instead run everything on the main thread. This has
been requested a couple of times and is useful as several things only
work on the main thread (OpenGL initialization, ...).