Deemphasize MSRV by removing check and note from README
CI got broken by `clap` releasing `4.1.0` that bumped its MRSV to 1.64.
Reading up on this again, I'm not sure how maintainers should best
communicate MSRV for crates that have dependencies. I think saying
"MSRV 1.60" in the README sets the expectation that this whole crate
works with 1.60, which might not be the case due to dependencies.

So... I left it in `Cargo.toml`, but otherwise this is not checked
anymore.