Include the kind of test while matching filters
Consider a trial with name "my_test" and kind "foo". The text that gets
printed out with `--list --format terse` is:

```
[foo] my_test: test
```

With libtest, an invariant is that the part of the string before ":
test" can be passed into "--exact" to run that test. However, that
invariant is currently not upheld by libtest-mimic: it only matches
against the name of the test and not the kind.

Make it so that libtest-mimic also matches against the kind of test, not
just its name.

I also took the liberty of running rustfmt (or really, my editor
automatically ran it when I went in to edit the file.)