Make `Arguments::is_filtered_out` public
Some test binaries want to offer extensions to the `libtest` interface
that operate on a set of tests. For example, a test binary may want to
allow the user to request the set of environmental dependencies (e.g.,
external artifacts to download) for running a set of tests. In such
cases, it is desirable to allow the user to specify test filters using
the same interface for filtering tests to run.

To this end, expose the existing `Arguments::is_filtered_out` function
publicly. This allows the test binary to apply the same filtering logic
without having to call and somehow extract the results from `run`.